pyproject.toml
[project]
name = "nba-elt-rest-api"
version = "1.9.0"
description = "REST API Project"
authors = [{ name = "jyablonski9", email = "jyablonski9@gmail.com" }]
requires-python = "~=3.12"
readme = "README.md"
dependencies = [
    "fastapi[all]>=0.103.2,<0.104",
    "sqlalchemy>=2.0.21,<3",
    "pydantic>=2.4.2,<3",
    "psycopg2-binary>=2.9.8,<3",
    "mangum>=0.17.0,<0.18",
    "jinja2>=3.1.2,<4",
    "opentelemetry-instrumentation-fastapi>=0.41b0,<0.42",
    "opentelemetry-instrumentation-requests>=0.41b0,<0.42",
    "opentelemetry-instrumentation-sqlalchemy>=0.41b0,<0.42",
    "opentelemetry-exporter-otlp>=1.20.0,<2",
    "opentelemetry-api>=1.20.0,<2",
    "opentelemetry-sdk>=1.20.0,<2",
    "protobuf>=4.24.3,<5",
    "python-jose>=3.3.0,<4",
    "pyyaml>=6.0.1,<7",
    "fastapi-cache2[redis]>=0.2.1,<0.3",
    "boto3>=1.34.41,<2",
    "authlib>=1.3.0,<2",
    "bcrypt==3.2.2",
    "cryptography==3.4.8",
    "pytz~=2024.1",
]

[project.urls]
Repository = "https://github.com/jyablonski/nba_elt_rest_api"

[dependency-groups]
dev = [
    "pytest>=7.4.2,<8",
    "pytest-mock>=3.11.1,<4",
    "pytest-cov>=4.1.0,<5",
    "coveralls>=3.3.1,<4",
    "prometheus-fastapi-instrumentator>=6.1.0,<7",
    "ruff>=0.1.2,<0.2",
    "types-pyyaml>=6.0.12.12,<7",
    "sqlalchemy-stubs>=0.4,<0.5",
    "types-python-jose>=3.3.4.8,<4",
    "pytest-asyncio>=0.21.1,<0.22",
    "boto3-stubs[lambda]>=1.34.62,<2",
]
local_dev = [
    "ipykernel>=6.25.2,<7",
    "mypy>=1.7.0,<2",
]

[tool.uv]
package = false
default-groups = [
    "dev",
    "local_dev",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
files = ["."]
plugins = "sqlmypy"

[[tool.mypy.overrides]]
module = ["src/database.py", "secondmodule", "utils", "config"]
ignore_missing_imports = true

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
    'ignore:function ham\(\) is deprecated:DeprecationWarning',
]
testpaths = ["tests"]
markers = [
    "wip: work in progress",
    "infrastructure: mocks for dummy data (sql db, http endpoint etc)",
]
addopts = "-v"
